minor auto adjustments, trying to make 2 coral work + grabber right trigger remap

+++ src/main/java/frc/robot/subsystems/GrabberSubsystem.java
@@ -45,7 +45,7 @@ public class GrabberSubsystem extends SubsystemBase{
     }
 
     //grab command we are actively using. just turns inward very fast
-    public Command activeIntake(int speed){
+    public Command activeIntake(double speed){
         final VoltageOut m_request = new VoltageOut(speed);
         return run(() -> {
             grabberMotor.setControl(m_request);
